delete_charater

+++ services/CharacterService/CharacterService.cs
@@ -90,10 +90,19 @@ namespace dotnet_rpg.services.CharacterService
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
-                serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                Character character = await _context.Characters
+                    .FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+                    if (character != null) 
+                    {
+                        _context.Characters.Remove(character);
+                        await _context.SaveChangesAsync();
+                        serviceResponse.Data = _context.Characters
+                            .Where(c => c.User.Id == GetUserId())
+                            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+                    } else {
+                        serviceResponse.Success = false;
+                        serviceResponse.Message = "Character not found";
+                    }
             }
             catch (Exception ex)
             {
